Move to the RC1 of commons-fileupload. This should be available on ibiblio shortly. If your compilation breaks, get it directly from jakarta.apache.org/commons/fileupload/ and put it into your local repo.
PR:
Obtained from:
Submitted by:
Reviewed by:


git-svn-id: https://svn.apache.org/repos/asf/jakarta/turbine/core/trunk@221482 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/turbine/services/upload/TurbineUpload.java
@@ -135,7 +135,7 @@ public abstract class TurbineUpload
      * @return The value of 'size.max' property of {@link
      * UploadService}.
      */
-    public static int getSizeMax()
+    public static long getSizeMax()
     {
         return getService().getSizeMax();
     }
